Restructure final loop of distributeCorest
The loop that actually assigns cores to threads now iterates over
threadsToReceiveCores instead of managedCores.

This is the first step in allowing threads to prefer certain cores over others,
for the purposes of keeping threads of the same application on hypertwins of
the same physical core.

The original code iterates over cores and puts a thread on them, picking the
threads in priority order.  If we continued to iterate over cores and pick
threads to run on them, we may violate priority order.

To preserve priority order while introducing affinity for cores, we must
instead iterate over threads and find a core to put them on. Later commits can
change the mechanism for finding a core without violating priority correctness.