Reopen cpuset tasks file when an error occurs on a thread read out of it.
Prior to this commit, if a thread exited because an application terminated, the
Core Arbiter would continue to read the thread's TID from the coreToCpusetFile
(cpuset tasks file), and continue trying (and failing) to write it to the tasks
file of the unmanaged cpuset. If we reopen the coreToCpusetFile, then we cease
to read the tids of deceased threads from it.